renamed modules to match convention
