--- Bridges/__init__.py
@@ -1,23 +1,0 @@
-from Bridges.data_src_dependent.DataSource import *
-from Bridges.Array import *
-from Bridges.AVLTreeElement import *
-from Bridges.BSTElement import *
-from Bridges.Bridges import *
-from Bridges.BSTElement import *
-from Bridges.CircDLelement import *
-from Bridges.CircSLelement import *
-from Bridges.Color import *
-from Bridges.Connector import *
-from Bridges.DLelement import *
-from Bridges.Edge import *
-from Bridges.Element import *
-from Bridges.ElementVisualizer import *
-from Bridges.GraphAdjList import *
-from Bridges.GraphAdjMatrix import *
-from Bridges.Grid import *
-from Bridges.LinkVisualizer import *
-from Bridges.MLelement import *
-from Bridges.SLelement import *
-from Bridges.Student import *
-from Bridges.TreeElement import *
-from Bridges.ColorGrid import *
