--- Bridges/data_src_dependent/__init__.py
@@ -1,1 +1,0 @@
-from Bridges.data_src_dependent.DataSource import *
