+++ bridges/sl_element.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-from Bridges.Element import *
+from bridges.element import *
 
 ##
 # 	@brief This class can be used to instantiate Singly Linked Elements.
@@ -33,7 +33,7 @@ class SLelement(Element):
     ##
     # This constructor creates an SLelement object
     # and sets the next pointer to null
-    # @param label the label of SLelement that shows up on the Bridges visualization
+    # @param label the label of SLelement that shows up on the bridges visualization
     # @param e the generic object that this SLelement will hold
     # @param next the element that should be assigned to the next pointer
     #
